Refactor, add a few docs
